+++ Treehouse/Treehouse/Network/Register/DTO/Response/PostAcceptInvitationTreeMemberResponseDTO.swift
@@ -10,4 +10,8 @@ import Foundation
 struct PostAcceptInvitationTreeMemberResponseDTO: Decodable {
     let treehouseId: Int
     let isAccepted: Bool
+    
+    func toDomain() -> AcceptInvitationTreeMemberResponseEntity {
+        return AcceptInvitationTreeMemberResponseEntity(treehouseId: treehouseId, isAccepted: isAccepted)
+    }
 }
